refactor: user 도메인 서비스 인터페이스 추상화하도록 리팩토링
#51
